// src/mocks/enums/image.ts
export enum Image {
  Apartment1 = '../../markup/img/apartment-01.jpg',
  Apartment2 = '../../markup/img/apartment-02.jpg',
  Apartment3 = '../../markup/img/apartment-03.jpg',
  Apartment3Small = '../../markup/img/apartment-small-03.jpg',
  Apartment4Small = '../../markup/img/apartment-small-04.jpg',
  Room = '../../markup/img/room.jpg',
  RoomSmall = '../../markup/img/room-small.jpg',
  Studio = '../../markup/img/studio-01.jpg',
  StudioPhotos = '../../markup/img/studio-photos.jpg',
  AvatarAngelina = '../../markup/img/avatar-angelina.jpg',
  AvatarMax = '../../markup/img/avatar-max.jpg'
}